colab_models/requirements_pro.txt
# Formula 1 Tire Change Prediction - Colab Pro Requirements
# Ottimizzato per Google Colab Pro con GPU e 25GB RAM

# Deep Learning Core
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
transformers>=4.30.0

# Data Science Stack
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Visualization & Monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
tensorboard>=2.13.0
wandb>=0.15.0
tqdm>=4.65.0

# Formula 1 Specific
fastf1>=3.1.0
pyarrow>=12.0.0

# Colab Integration
google-colab
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Web Framework (for API)
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0

# Memory & Performance
psutil>=5.9.0
memory_profiler>=0.60.0
numba>=0.57.0

# Data Handling
h5py>=3.9.0
joblib>=1.3.0
dill>=0.3.6

# Jupyter Enhancements
ipywidgets>=8.0.0
jupyter>=1.0.0
notebook>=6.5.0

# Machine Learning Extensions
optuna>=3.2.0
hyperopt>=0.2.7
imbalanced-learn>=0.11.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.0
rich>=13.4.0

# Development & Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
